Refactored WebSocket & Measurement
This change implements the required feature that a measurement can be started and stopped unrelated to a WebSocket client being connected.
This is very basic for now, as you have to manually reload the page when you start a measurement to get the option to conneect.
Also, the UI needs to be adapted to better support this.